Added Functions to automate character creation
GameController can now generate a random or selected character. Random
currently "just makes a warrior", although this can generate any random
job, just gotta toss a  Rand() and a switch statement and just toss
classes in as they come in.

To update the selectedCharacter, the same thing has to happen, but we
also need to add a line of COUT to present the choice.

Eventually, the COUT comments and CIN should map to our AddMessage and
getch() commands.