Movement Updates
Boulders can now be pushed, need to make some switches to test it.

Enemies can also push boulders.

Enemies and players need placed on maps randomly,
mapReader needs to be able to load random maps and switch to the new
input format, once Character.fight() is merged that will be tested
(placeholder ready for it).